commands - migrate pipeline command
- move legacy /routes to /need-revision before proper migration
